+++ backend/chainlit/server.py
@@ -6,6 +6,16 @@ import os
 import re
 import shutil
 import urllib.parse
+from typing import Any, Optional, Union
+
+from chainlit.oauth_providers.providers import get_oauth_provider
+from chainlit.secret import random_secret
+
+mimetypes.add_type("application/javascript", ".js")
+mimetypes.add_type("text/css", ".css")
+
+import asyncio
+import os
 import webbrowser
 from contextlib import asynccontextmanager
 from pathlib import Path
